Move Api entities to Core
