+++ src/Wallabag/CoreBundle/Controller/Api/WallabagRestController.php
@@ -16,7 +16,7 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInt
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Contracts\Translation\TranslatorInterface;
-use Wallabag\ApiBundle\Entity\ApplicationInfo;
+use Wallabag\CoreBundle\Entity\Api\ApplicationInfo;
 use Wallabag\CoreBundle\Entity\User;
 
 class WallabagRestController extends AbstractFOSRestController
